+++ src/dipdup/datasources/evm_node.py
@@ -46,6 +46,8 @@ NODE_LEVEL_TIMEOUT = 1.0
 # NOTE: This value was chosen empirically and likely is not optimal.
 NODE_BATCH_SIZE = 32
 NODE_LAST_MILE = 128
+POLL_INTERVAL = 5
+
 
 EmptyCallback = Callable[[], Awaitable[None]]
 HeadCallback = Callable[['EvmNodeDatasource', EvmNodeHeadData], Awaitable[None]]
@@ -84,7 +86,6 @@ class EvmNodeDatasource(IndexDatasource[EvmNodeDatasourceConfig]):
         super().__init__(config, merge_subscriptions)
         self._web3_client: AsyncWeb3 | None = None
         self._ws_client: WebsocketTransport | None = None
-        self._use_realtime: asyncio.Event = asyncio.Event()
         self._requests: dict[str, tuple[asyncio.Event, Any]] = {}
         self._subscription_ids: dict[str, EvmNodeSubscription] = {}
         self._emitter_queue: Queue[LevelData] = Queue()
@@ -131,12 +132,17 @@ class EvmNodeDatasource(IndexDatasource[EvmNodeDatasourceConfig]):
         self.set_sync_level(None, level)
 
     async def run(self) -> None:
-        await self._use_realtime.wait()
-        await asyncio.gather(
-            self._ws_loop(),
-            self._emitter_loop(),
-            self._watchdog.run(),
-        )
+        if self.realtime:
+            await asyncio.gather(
+                self._ws_loop(),
+                self._emitter_loop(),
+                self._watchdog.run(),
+            )
+        else:
+            while True:
+                level = await self.get_head_level()
+                self.set_sync_level(None, level)
+                await asyncio.sleep(POLL_INTERVAL)
 
     async def _emitter_loop(self) -> None:
         while True:
@@ -173,11 +179,12 @@ class EvmNodeDatasource(IndexDatasource[EvmNodeDatasourceConfig]):
 
         raise DatasourceError('Websocket connection failed', self.name)
 
-    def use_realtime(self) -> None:
-        self._use_realtime.set()
+    @property
+    def realtime(self) -> bool:
+        return self._config.ws_url is not None
 
     async def subscribe(self) -> None:
-        if not self._use_realtime.is_set():
+        if not self.realtime:
             return
 
         missing_subscriptions = self._subscriptions.missing_subscriptions
@@ -394,6 +401,8 @@ class EvmNodeDatasource(IndexDatasource[EvmNodeDatasourceConfig]):
         self._logger.debug('Creating Websocket client')
 
         url = self._config.ws_url
+        if not url:
+            raise FrameworkException('Spawning node datasource, but `ws_url` is not set')
         self._ws_client = WebsocketTransport(
             url=url,
             protocol=WebsocketProtocol(),